Jenkinsfile: Update the node labels
Since PR https://github.com/Microsoft/openenclave/pull/1248, the `hardware` label is deprecated.

Right now, we have `ACC-1604` and `ACC-1804` for a better separation of
the tested Ubuntu releases.

The previous `hardware` label will be converted to `ACC-1604`. In the future,
once Ubuntu 18.04 is validated, we'll be adding tests for `ACC-1804` as well.