Use Cache-Control header to determine how long collateral should be cached for (#122)
Allow PCK Cache to determine how long to cache collateral. Ordinarily, collateral should always be cached for 24 hours but PCK cache will examine the expiration time of the collateral and notify the Azure DCAP Client how long the collateral should be cached for.

For example, if the collateral will expire in 3 hours, the cache-control header will tell Azure DCAP Client to cache the collateral for only three hours so Azure DCAP client will re-query for newer collateral in that time frame.

This PR also addresses issue #114 by re-classifying errors reading environment variables to be warning instead of errors. The library is configured to operate with sensible defaults if no environment variables are defined.